byteos-intelligence/requirements.txt
# ByteOS Intelligence — Python Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
supabase==2.7.4
together==1.3.0
openai==1.50.0
anthropic==0.34.0
httpx==0.27.2
numpy==1.26.4
tiktoken==0.7.0
python-multipart==0.0.12
PyJWT>=2.8.0
edge-tts>=6.1.0
pydub>=0.25.0
